Moved log out button to navigation drawer
