+++ app/src/main/res/values/dimens.xml
@@ -59,4 +59,8 @@
 
     <dimen name="broadcast_card_view_min_height">100dp</dimen>
     <dimen name="text_view_two_lines_fixed_size">36dp</dimen>
+
+    <dimen name="fragment_impact_textview_no_width">0dp</dimen>
+    <dimen name="fragment_impact_textview_no_height">0dp</dimen>
+    <dimen name="fragment_impact_no_border_width">0dp</dimen>
 </resources>
